add test with mock fetch for svc
