+++ src/services/api/RetrieveParsedEquationsService.tsx
